# docs/site/content/en/openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: Horreum REST API
  description: "Horreum automated change anomaly detection. For more information,\
    \ please see [https://horreum.hyperfoil.io/](https://horreum.hyperfoil.io/)"
  version: "0.13"
...
paths:
...
  /api/dataset/bySchema:
    get:
      tags:
      - Dataset
      description: "Retrieve a paginated list of Datasets, with total count, by Schema"
      operationId: listDatasetsBySchema
      parameters:
      - name: uri
        in: query
        description: Schema URI
        required: true
        schema:
          type: string
        example: uri:techempower:0.1
      - name: limit
        in: query
        description: limit the number of results
        schema:
          format: int32
          type: integer
        example: 20
      - name: page
        in: query
        description: filter by page number of a paginated list of Schemas
        schema:
          format: int32
          type: integer
        example: 2
      - name: sort
        in: query
        description: Field name to sort results
        schema:
          default: start
          type: string
        example: name
      - name: direction
        in: query
        description: Sort direction
        schema:
          $ref: '#/components/schemas/SortDirection'
        example: Ascending
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetList'
  /api/dataset/list/{testId}:
    get:
      tags:
      - Dataset
      description: "Retrieve a paginated list of Datasets, with total count, by Test"
      operationId: listByTest
      parameters:
      - name: testId
        in: path
        description: Test ID of test to retrieve list of Datasets
        required: true
        schema:
          format: int32
          type: integer
        example: 101
      - name: filter
        in: query
        description: JOSN Filter expression to apply to query
        schema:
          type: string
        example:
          buildID: 111111
      - name: limit
        in: query
        description: limit the number of results
        schema:
          format: int32
          type: integer
        example: 20
      - name: page
        in: query
        description: filter by page number of a paginated list of Schemas
        schema:
          format: int32
          type: integer
        example: 2
      - name: sort
        in: query
        description: Field name to sort results
        schema:
          type: string
        example: name
      - name: direction
        in: query
        description: Sort direction
        schema:
          $ref: '#/components/schemas/SortDirection'
        example: Ascending
      - name: viewId
        in: query
        description: Optional View ID to filter datasets by view
        schema:
          format: int32
          type: integer
        example: 202
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetList'
  /api/dataset/{datasetId}/labelValues:
    get:
      tags:
      - Dataset
      operationId: datasetLabelValues
      parameters:
      - name: datasetId
        in: path
        required: true
        schema:
          format: int32
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LabelValue'
  /api/dataset/{datasetId}/previewLabel:
    post:
      tags:
      - Dataset
      operationId: previewLabel
      parameters:
      - name: datasetId
        in: path
        required: true
        schema:
          format: int32
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Label'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabelPreview'
  /api/dataset/{datasetId}/summary:
    get:
      tags:
      - Dataset
      operationId: getSummary
      parameters:
      - name: datasetId
        in: path
        required: true
        schema:
          format: int32
          type: integer
      - name: viewId
        in: query
        schema:
          format: int32
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetSummary'
  /api/dataset/{id}:
    get:
      tags:
      - Dataset
      description: Retrieve Dataset by ID
      operationId: getDataset
      parameters:
      - name: id
        in: path
        description: Dataset ID to retrieve
        required: true
        schema:
          format: int32
          type: integer
        example: 101
      responses:
        "404":
          description: No Dataset with the given id was found
          content:
            application/json: {}
        "200":
          description: JVM system properties of a particular host.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dataset'
...
  /api/test/summary:
    get:
      tags:
      - Test
      description: Retrieve a summary of Tests in a folder
      operationId: summary
      parameters:
      - name: roles
        in: query
        description: "\"__my\", \"__all\" or a comma delimited  list of roles"
        schema:
          type: string
        example: __my
      - name: folder
        in: query
        description: name of the Folder containing the Tests
        schema:
          type: string
        example: My Team Folder
      - name: limit
        in: query
        description: limit the result count
        schema:
          default: 20
          type: integer
        example: 20
      - name: page
        in: query
        description: 'filter by page number of a paginated list of '
        schema:
          default: 1
          type: integer
        example: 1
      - name: direction
        in: query
        description: Sort direction
        schema:
          allOf:
          - $ref: '#/components/schemas/SortDirection'
          - default: Ascending
        example: Ascending
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestListing'
...
components:
  schemas:
    Access:
      description: "Resources have different visibility within the UI. 'PUBLIC', 'PROTECTED'\
        \ and 'PRIVATE'. Restricted resources are not visible to users who do not\
        \ have the correct permissions"
      enum:
      - PUBLIC
      - PROTECTED
      - PRIVATE
      type: string
...
    Dataset:
      description: A dataset is the JSON document used as the basis for all comparisons
        and reporting
      required:
      - testid
      - data
      - ordinal
      type: object
      allOf:
      - $ref: '#/components/schemas/ProtectedType'
      - $ref: '#/components/schemas/ProtectedTimeType'
      properties:
        id:
          format: int32
          description: Dataset Unique ID
          type: integer
          example: 101
        description:
          description: Run description
          type: string
          example: Run on AWS with m7g.large
        testid:
          format: int32
          description: Test ID that Dataset relates to
          type: integer
          example: 101
        data:
          description: Data payload
          type: string
        ordinal:
          format: int32
          description: Dataset ordinal for ordered list of Datasets derived from a
            Run
          type: integer
          example: 1
        validationErrors:
          description: List of Validation Errors
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
        runId:
          format: int32
          description: Run ID that Dataset relates to
          type: integer
          example: 101
...
    DatasetList:
      description: Result containing a subset of Dataset Summaries and the total count
        of available. Used in paginated tables
      required:
      - total
      - datasets
      type: object
      properties:
        total:
          format: int64
          description: Total number of Dataset Summaries available
          type: integer
          example: 64
        datasets:
          description: List of Dataset Summaries. This is often a subset of total
            available.
          type: array
          items:
            $ref: '#/components/schemas/DatasetSummary'
...
    DatasetSummary:
      required:
      - id
      - runId
      - ordinal
      - testId
      - testname
      - schemas
      type: object
      allOf:
      - $ref: '#/components/schemas/ProtectedTimeType'
      properties:
        id:
          format: int32
          description: Unique Dataset ID
          type: integer
          example: 101
        runId:
          format: int32
          description: Run ID that Dataset relates to
          type: integer
          example: 202
        ordinal:
          format: int32
          description: Ordinal position of Dataset Summary on returned List
          type: integer
          example: 3
        testId:
          format: int32
          description: Test ID that Dataset relates to
          type: integer
          example: 202
        testname:
          description: Test name that the Dataset relates to
          type: string
          example: my-comprehensive-benchmark
        description:
          description: Dataset description
          type: string
          example: Run on AWS with m7g.large
        view:
          description: map of view component ids to the LabelValueMap to render the
            component for this dataset
          type: object
          allOf:
          - $ref: '#/components/schemas/IndexedLabelValueMap'
          example: "{ \"[view_component_id]\": { \"[labelName]\": labelValue} }"
        schemas:
          description: List of Schema usages
          type: array
          items:
            $ref: '#/components/schemas/SchemaUsage'
        validationErrors:
          description: List of Validation Errors
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
...
    ErrorDetails:
      required:
      - type
      - message
      type: object
      properties:
        type:
          description: Validation Error type
          type: string
        code:
          type: string
        path:
          type: string
        evaluationPath:
          type: string
        schemaPath:
          type: string
          deprecated: true
        schemaLocation:
          type: string
        instanceLocation:
          type: string
        property:
          type: string
        arguments:
          type: array
          items:
            type: string
        details:
          type: string
        messageKey:
          type: string
        valid:
          type: boolean
        message:
          type: string
...
    Extractor:
      description: "An Extractor defines how values are extracted from a JSON document,\
        \ for use in Labels etc."
      required:
      - name
      - jsonpath
      - isarray
      type: object
      properties:
        name:
          description: Name of extractor. This name is used in Combination Functions
            to refer to values by name
          type: string
          example: buildID
        jsonpath:
          description: JSON path expression defining the location of the extractor
            value in the JSON document. This is a pSQL json path expression
          type: string
          example: $.buildInfo.buildID
        isarray:
          description: Does the JSON path expression reference an Array?
          type: boolean
          example: false
...
    IndexedLabelValueMap:
      type: object
      additionalProperties:
        $ref: '#/components/schemas/LabelValueMap'
...
    Label:
      description: "A Label is a core component of Horreum, defining which components\
        \ of the JSON document are part of a KPI and how the metric values are calculated"
      required:
      - id
      - name
      - extractors
      - filtering
      - metrics
      - schemaId
      type: object
      allOf:
      - $ref: '#/components/schemas/ProtectedType'
      properties:
        id:
          format: int32
          description: Unique ID for Label
          type: integer
          example: 101
        name:
          description: "Name for label. NOTE: all Labels are considered to have the\
            \ same semantic meaning throughout the entire system"
          type: string
          example: Throughput
        extractors:
          description: "A collection of Extractors, that will be combined in the Combination\
            \ Function"
          type: array
          items:
            $ref: '#/components/schemas/Extractor'
        function:
          description: A Combination Function that defines how values from Extractors
            are combined to produce a Label Value
          type: string
          example: "value => { return ((value.reduce((a,b) => a+b))/value.length*1000).toFixed(3);\
            \ }"
        filtering:
          description: Is Label a filtering label? Filtering labels contains values
            that are used to filter datasets for comparison
          type: boolean
          example: true
        metrics:
          description: Is Label a metrics label? Metrics labels are contain Metrics
            that are used for comparison
          type: boolean
          example: true
        schemaId:
          format: int32
          description: Schema ID that the Label relates to
          type: integer
          example: 101
...
    LabelPreview:
      description: Preview a Label Value derived from a Dataset Data. A preview allows
        users to apply a Label to a dataset and preview the Label Value result and
        processing errors in the UI
      type: object
      properties:
        value:
          description: "Value value extracted from Dataset. This can be a scalar,\
            \ array or JSON object"
          type: string
        output:
          description: Description of errors occurred attempting to generate Label
            Value Preview
          type: string
    LabelValue:
      description: Label Value derived from Label definition and Dataset Data
      required:
      - id
      - name
      - schema
      type: object
      properties:
        id:
          format: int32
          description: Unique ID for Label Value
          type: integer
          example: 101
        name:
          description: Label name
          type: string
          example: buildID
        schema:
          description: Summary description of Schema
          type: object
          allOf:
          - $ref: '#/components/schemas/SchemaDescriptor'
        value:
          description: "Value value extracted from Dataset. This can be a scalar,\
            \ array or JSON object"
          type: string
          example: "1724"
    LabelValueMap:
      description: a map of label name to value
      type: object
      example: "{ \"[labelName]\": labelValue}"
...
    ProtectedTimeType:
      required:
      - access
      - owner
      - start
      - stop
      type: object
      properties:
        access:
          description: Access rights for the test. This defines the visibility of
            the Test in the UI
          enum:
          - PUBLIC
          - PROTECTED
          - PRIVATE
          type: string
          allOf:
          - $ref: '#/components/schemas/Access'
          example: PUBLIC
        owner:
          description: Name of the team that owns the test. Users must belong to the
            team that owns a test to make modifications
          type: string
          example: performance-team
        start:
          format: int64
          description: Run Start timestamp
          type: integer
          example: 1704965908267
        stop:
          format: int64
          description: Run Stop timestamp
          type: integer
          example: 1704965908267
    ProtectedType:
      required:
      - access
      - owner
      type: object
      properties:
        access:
          description: Access rights for the test. This defines the visibility of
            the Test in the UI
          enum:
          - PUBLIC
          - PROTECTED
          - PRIVATE
          type: string
          allOf:
          - $ref: '#/components/schemas/Access'
          example: PUBLIC
        owner:
          description: Name of the team that owns the test. Users must belong to the
            team that owns a test to make modifications
          type: string
          example: performance-team
...
    SchemaDescriptor:
      required:
      - id
      - name
      - uri
      type: object
      properties:
        id:
          format: int32
          description: Schema unique ID
          type: integer
          example: 1
        name:
          description: Schema name
          type: string
          example: my-benchmark-schema
        uri:
          description: Schema name
          type: string
          example: uri:my-schmea:0.1
...
    SchemaUsage:
      required:
      - id
      - name
      - uri
      - source
      - type
      - hasJsonSchema
      type: object
      properties:
        id:
          format: int32
          description: Schema unique ID
          type: integer
          example: 1
        name:
          description: Schema name
          type: string
          example: my-benchmark-schema
        uri:
          description: Schema name
          type: string
          example: uri:my-schmea:0.1
        source:
          format: int32
          description: "Source of schema usage, 0 is data, 1 is metadata. DataSets\
            \ always use 0"
          type: integer
          example: 1
        type:
          format: int32
          description: "Location of Schema Usage, 0 for Run, 1 for Dataset"
          type: integer
          example: 1
        key:
          description: Ordinal position of schema usage in Run/Dataset
          type: string
          example: "1"
        hasJsonSchema:
          description: Does schema have a JSON validation schema defined?
          type: boolean
          example: false
...
    SortDirection:
      enum:
      - Ascending
      - Descending
      type: string
...
    TestListing:
      required:
      - tests
      - count
      type: object
      properties:
        tests:
          description: Array of Test Summaries
          type: array
          items:
            $ref: '#/components/schemas/TestSummary'
        count:
          format: int64
          description: Number of tests when pagination is ignored
          type: integer
...
    TestSummary:
      required:
      - id
      - name
      - datastoreId
      type: object
      allOf:
      - $ref: '#/components/schemas/ProtectedType'
      properties:
        id:
          format: int32
          description: ID of tests
          type: integer
          example: 101
        name:
          description: Test name
          type: string
          example: my-comprehensive-benchmark
        folder:
          description: Name of folder that the test is stored in. Folders allow tests
            to be organised in the UI
          type: string
          example: My Team Folder
        description:
          description: Description of the test
          type: string
          example: Comprehensive benchmark to tests the limits of any system it is
            run against
        datasets:
          description: Total number of Datasets for the Test
          type: number
          example: 202
        runs:
          description: Total number of Runs for the Test
          type: number
          example: 101
        watching:
          description: Subscriptions for each test for authenticated user
          uniqueItems: true
          type: array
          items:
            type: string
          example: []
        datastoreId:
          format: int32
          description: Datastore id
          type: integer
          example: 1
...
    ValidationError:
      required:
      - schemaId
      - error
      type: object
      properties:
        schemaId:
          format: int32
          description: Schema ID that Validation Error relates to
          type: integer
          example: 101
        error:
          description: Validation Error Details
          type: object
          allOf:
          - $ref: '#/components/schemas/ErrorDetails'
...